@@ -38,35 +38,37 @@
   ],
   "dependencies": {
     "assets-webpack-plugin": "^3.4.0",
-    "autoprefixer": "^6.3.7",
-    "babel-core": "^6.11.4",
+    "autoprefixer": "^6.4.0",
+    "babel-core": "^6.14.0",
     "babel-eslint": "^6.1.2",
-    "babel-loader": "^6.2.4",
+    "babel-loader": "^6.2.5",
     "babel-plugin-transform-decorators": "^6.8.0",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-runtime": "^6.12.0",
-    "babel-preset-es2015": "^6.9.0",
+    "babel-preset-es2015": "^6.14.0",
     "babel-preset-react": "^6.11.1",
     "babel-preset-stage-0": "^6.5.0",
-    "babel-preset-stage-1": "^6.5.0",
-    "babel-register": "^6.11.6",
+    "babel-preset-stage-1": "^6.13.0",
+    "babel-register": "^6.14.0",
     "babel-root-import": "^4.0.1",
     "babel-runtime": "^6.11.6",
+    "bluebird": "^3.4.3",
     "browser-sync": "^2.14.0",
     "chai": "^3.5.0",
     "chalk": "^1.1.3",
     "copy-webpack-plugin": "^3.0.1",
     "cpy": "^4.0.1",
     "css-loader": "^0.23.1",
-    "cssnano": "^3.7.3",
+    "cssnano": "^3.7.4",
     "del": "^2.2.1",
     "dotenv": "^2.0.0",
     "ejs": "^2.5.1",
-    "eslint": "^3.2.2",
-    "eslint-config-airbnb": "^9.0.1",
-    "eslint-plugin-import": "^1.12.0",
-    "eslint-plugin-jsx-a11y": "^2.0.1",
-    "eslint-plugin-react": "^5.2.2",
+    "eslint": "^3.3.1",
+    "eslint-config-airbnb": "^10.0.0",
+    "eslint-plugin-import": "^1.14.0",
+    "eslint-plugin-jsx-a11y": "^2.1.0",
+    "eslint-plugin-jsx-control-statements": "^2.1.1",
+    "eslint-plugin-react": "^6.1.2",
     "file-loader": "^0.9.0",
     "font-awesome": "^4.6.3",
     "font-awesome-webpack": "0.0.4",
@@ -77,9 +79,9 @@
     "json-stringify-safe": "^5.0.1",
     "jsx-control-statements": "^3.1.2",
     "less": "^2.7.1",
-    "markdown-it": "^7.0.0",
+    "markdown-it": "^7.0.1",
     "path-to-regexp": "^1.5.3",
-    "pixrem": "^3.0.1",
+    "pixrem": "^3.0.2",
     "pleeease-filters": "^3.0.0",
     "postcss": "^5.1.1",
     "postcss-calc": "^5.3.0",
@@ -96,10 +98,10 @@
     "react-hot-loader": "^3.0.0-beta.2",
     "simple-universal-style-loader": "^0.14.4",
     "style-loader": "^0.13.1",
-    "stylelint": "^7.0.3",
-    "stylelint-config-standard": "^11.0.0",
+    "stylelint": "^7.1.0",
+    "stylelint-config-standard": "^12.0.0",
     "url-loader": "^0.5.7",
-    "webpack": "^1.13.1",
+    "webpack": "^1.13.2",
     "webpack-dev-middleware": "^1.6.1",
     "webpack-hot-middleware": "^2.12.2",
     "webpack-middleware": "^1.5.1"
@@ -108,11 +110,13 @@
     "react-app": "^1.1.2"
   },
   "devDependencies": {
+    "babel-plugin-transform-promise-to-bluebird": "^1.1.0",
+    "eslint-loader": "^1.5.0",
     "history": "^3.0.0",
-    "react": "^15.3.0",
-    "react-dom": "^15.3.0",
-    "mobx": "^2.4.1",
-    "mobx-react": "^3.5.3"
+    "mobx": "^2.4.4",
+    "mobx-react": "^3.5.5",
+    "react": "^15.3.1",
+    "react-dom": "^15.3.1"
   },
   "babel": {
     "presets": [
@@ -129,7 +133,18 @@
   },
   "eslintConfig": {
     "parser": "babel-eslint",
-    "extends": "airbnb"
+    "extends": "airbnb",
+    "rules": {
+      "react/jsx-filename-extension": [
+        1,
+        {
+          "extensions": [
+            ".js",
+            ".jsx"
+          ]
+        }
+      ]
+    }
   },
   "stylelint": {
     "extends": "stylelint-config-standard",
