@@ -5,15 +5,15 @@ function Footer() {
     return (
         <div className={s.root}>
             <div className={s.container}>
-                <span className={s.text}>© Your Company</span>
+                <span className={s.text}>© SkiScool</span>
                 <span className={s.spacer}>·</span>
                 <Link className={s.link} to="Home">Home</Link>
                 <span className={s.spacer}>·</span>
-                <Link className={s.link} to="root">root</Link>
+                <Link className={s.link} to="Root">root</Link>
                 <span className={s.spacer}>·</span>
-                <Link className={s.link} to="/privacy">Privacy</Link>
+                <Link className={s.link} to="Privacy">Privacy</Link>
                 <span className={s.spacer}>·</span>
-                <Link className={s.link} to="/not-found">Not Found</Link>
+                <Link className={s.link} to="Not">Not Found</Link>
             </div>
         </div>
     );
