@@ -8,11 +8,11 @@ function Navigation({className}) {
     return (
         <div className={cx(s.root)} role="navigation">
             <Link className={cx(s.link)} to="GetStarted" hoverColor="#8AA62F" icon={<i className="mb3 fa fa-diamond" />}>About</Link>
-            <Link className={cx(s.link,className)} to="contact">Contact</Link>
-            <Link className={cx(s.link,className)} to="root">Root</Link>
+            <Link className={cx(s.link,className)} to="Contact">Contact</Link>
+            <Link className={cx(s.link,className)} to="Root">Root</Link>
             <Link className={cx(s.link,className)} to="Profile">profile</Link>
             <span className={cx(s.spacer,className)}> | </span>
-            <Link className={cx(s.link, s.highlight,className)} to="/test">Test 404</Link>
+            <Link className={cx(s.link, s.highlight,className)} to="test">Test 404</Link>
         </div>
     );
 }
