@@ -3,12 +3,12 @@ import s from './GetStarted.css';
 import { title, html } from './GetStarted.md';
 
 class AboutPage extends React.Component {
-
-  componentDidMount() {
-    document.title = title;
-  }
+ static contextTypes = {
+setTitle: React.PropTypes.func.isRequired
+}
 
   render() {
+  this.context.setTitle(title)
     return (
       <Layout className={s.content}>
         <h1>{title}</h1>
