@@ -1,20 +1,20 @@
-import { lightWhite } from 'material-ui/styles/colors';
+//import { lightWhite } from 'material-ui/styles/colors';
 
 /*
   material-ui override styles
 */
 export default {
- /* palette: {
+palette: {
     primary1Color: '#5BC0BE',
     textColor: '#f0f0f0',
   },
-  textField: {
+ /* textField: {
     errorColor: '#F25F5C',
   },
   overlay: {
     backgroundColor: lightWhite,
-  },
+  },*/
    drawer: {
     color: '#1C2541', // this is the background-color
-  },*/
+  },
 };
